Refactor ISSN-registry request list and item views
