+++ src/frontend/components/issn-registry/subComponents/modals/IssnPublicationCreationModal.jsx
@@ -27,6 +27,10 @@
 
 import React, {useState} from 'react';
 import PropTypes from 'prop-types';
+
+import {useAuth} from 'react-oidc-context';
+import {useHistory} from 'react-router-dom';
+
 import {FormattedMessage} from 'react-intl';
 
 import {
@@ -41,6 +45,8 @@ import {
   Typography
 } from '@mui/material';
 
+import useAppStateDispatch from '/src/frontend/hooks/useAppStateDispatch';
+
 import {makeApiRequest} from '/src/frontend/actions';
 import {redirect} from '/src/frontend/actions/util';
 import {publishingLanguages, mediumOptions} from '/src/frontend/components/common/form/constants';
@@ -48,9 +54,12 @@ import {publishingLanguages, mediumOptions} from '/src/frontend/components/commo
 import '/src/frontend/css/common.css';
 import '/src/frontend/css/subComponents/modals.css';
 
-const IssnPublicationCreationModal = (props) => {
-  const {formId, userInfo, history, setSnackbarMessage} = props;
-  const {authenticationToken} = userInfo;
+const IssnPublicationCreationModal = ({formId}) => {
+  const history = useHistory();
+  const {user: {access_token: authenticationToken}} = useAuth();
+
+  const appStateDispatch = useAppStateDispatch();
+  const setSnackbarMessage = (snackbarMessage) => appStateDispatch({snackbarMessage});
 
   // State for the modal window
   const [openCreatePublicationModal, setOpenCreatePublicationModal] = useState(false);
@@ -238,10 +247,7 @@ const IssnPublicationCreationModal = (props) => {
 };
 
 IssnPublicationCreationModal.propTypes = {
-  formId: PropTypes.number.isRequired,
-  history: PropTypes.object.isRequired,
-  userInfo: PropTypes.object.isRequired,
-  setSnackbarMessage: PropTypes.func.isRequired
+  formId: PropTypes.number.isRequired
 };
 
 export default IssnPublicationCreationModal;
