+++ src/frontend/components/issn-registry/subComponents/modals/IssnRequestArchiveModal.jsx
@@ -27,6 +27,9 @@
 
 import React, {useState} from 'react';
 import PropTypes from 'prop-types';
+
+import {useAuth} from 'react-oidc-context';
+
 import {FormattedMessage} from 'react-intl';
 import moment from 'moment';
 import {Modal, Box, Typography, Button} from '@mui/material';
@@ -37,8 +40,8 @@ import '/src/frontend/css/common.css';
 import '/src/frontend/css/subComponents/modals.css';
 import Spinner from '/src/frontend/components/common/Spinner.jsx';
 
-function IssnRequestArchiveModal({formId, userInfo}) {
-  const {authenticationToken} = userInfo;
+function IssnRequestArchiveModal({formId}) {
+  const {user: {access_token: authenticationToken}} = useAuth();
 
   const [isModalOpen, setIsModalOpen] = useState(false); // State of the modal window (open/closed)
 
@@ -55,114 +58,6 @@ function IssnRequestArchiveModal({formId, userInfo}) {
     isModalOpen
   });
 
-  const component = getComponent();
-
-  function getComponent() {
-    if (loading) {
-      return <Spinner />;
-    }
-
-    if (error) {
-      return <Typography>Could not fetch data due to API error</Typography>;
-    }
-
-    return (
-      <>
-        <Typography variant="h5">
-          <FormattedMessage id="modal.issn.archive.record" />
-        </Typography>
-        <div className="archiveRecordContainer">
-          <div>
-            <Typography>
-              <FormattedMessage id="common.publisher.issn" />:
-            </Typography>
-            <p>{data.publisher}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="modal.issn.archive.amount" />:
-            </Typography>{' '}
-            <p>{data.publicationCount}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="form.common.contactPerson" />:
-            </Typography>{' '}
-            <p>{data.contactPerson}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="form.common.address" />:
-            </Typography>{' '}
-            <p>{data.address}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="form.common.zip" />:
-            </Typography>{' '}
-            <p>{data.zip}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="form.common.city" />:
-            </Typography>{' '}
-            <p>{data.city}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="form.common.email" />:
-            </Typography>{' '}
-            <p>{data.email}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="form.common.phone" />:
-            </Typography>{' '}
-            <p>{data.phone}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="form.common.language" />:
-            </Typography>{' '}
-            <p>
-              <FormattedMessage id={`common.${data.langCode}`} />
-            </p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="common.request" />:
-            </Typography>{' '}
-            <p>{data.formId}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="modal.issn.archive.id" />:
-            </Typography>{' '}
-            <p>{data.id}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="modal.issn.archive.oldId" />:
-            </Typography>{' '}
-            <p>{data.oldId ?? '-'}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="form.common.created" />:
-            </Typography>{' '}
-            <p>{moment(data.created).format('LLL')}</p>
-          </div>
-          <div>
-            <Typography>
-              <FormattedMessage id="form.common.createdBy" />:
-            </Typography>{' '}
-            <p>{data.createdBy}</p>
-          </div>
-        </div>
-      </>
-    );
-  }
-
   return (
     <>
       <Button
@@ -174,15 +69,116 @@ function IssnRequestArchiveModal({formId, userInfo}) {
         <FormattedMessage id="modal.issn.archive.record" />
       </Button>
       <Modal open={isModalOpen} onClose={() => setIsModalOpen(false)}>
-        <Box className="marcModal">{component}</Box>
+        <Box className="marcModal">
+          {/* Loading spinner */}
+          {loading && <Spinner />}
+
+          {/* Errors */}
+          {error && <Typography>Could not fetch data due to API error</Typography>}
+
+          {/* Content */}
+          {!loading && !error && (
+            <>
+              <Typography variant="h5">
+                <FormattedMessage id="modal.issn.archive.record" />
+              </Typography>
+              <div className="archiveRecordContainer">
+                <div>
+                  <Typography>
+                    <FormattedMessage id="common.publisher.issn" />:
+                  </Typography>
+                  <p>{data.publisher}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="modal.issn.archive.amount" />:
+                  </Typography>{' '}
+                  <p>{data.publicationCount}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="form.common.contactPerson" />:
+                  </Typography>{' '}
+                  <p>{data.contactPerson}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="form.common.address" />:
+                  </Typography>{' '}
+                  <p>{data.address}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="form.common.zip" />:
+                  </Typography>{' '}
+                  <p>{data.zip}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="form.common.city" />:
+                  </Typography>{' '}
+                  <p>{data.city}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="form.common.email" />:
+                  </Typography>{' '}
+                  <p>{data.email}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="form.common.phone" />:
+                  </Typography>{' '}
+                  <p>{data.phone}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="form.common.language" />:
+                  </Typography>{' '}
+                  <p>
+                    <FormattedMessage id={`common.${data.langCode}`} />
+                  </p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="common.request" />:
+                  </Typography>{' '}
+                  <p>{data.formId}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="modal.issn.archive.id" />:
+                  </Typography>{' '}
+                  <p>{data.id}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="modal.issn.archive.oldId" />:
+                  </Typography>{' '}
+                  <p>{data.oldId ?? '-'}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="form.common.created" />:
+                  </Typography>{' '}
+                  <p>{moment(data.created).format('LLL')}</p>
+                </div>
+                <div>
+                  <Typography>
+                    <FormattedMessage id="form.common.createdBy" />:
+                  </Typography>{' '}
+                  <p>{data.createdBy}</p>
+                </div>
+              </div>
+            </>)}
+        </Box>
       </Modal>
     </>
   );
 }
 
 IssnRequestArchiveModal.propTypes = {
-  formId: PropTypes.number.isRequired,
-  userInfo: PropTypes.object.isRequired
+  formId: PropTypes.number.isRequired
 };
 
 export default IssnRequestArchiveModal;
